Refactor BranchEntityPostgresWriterGrain to use IBranchWriter for database operations and add BranchEntityPostgresWriter for improved separation of concerns
